1st round of compilation fixes for mingw
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@2977 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/modules/iograph/libkviiograph.cpp
@@ -187,16 +187,16 @@ void KviIOGraphWidget::paintEvent(QPaintEvent *)
 
 	for(int i=0;i<=KVI_IOGRAPH_HORIZ_SEGMENTS;i++)
 	{
-		p.drawLine(0, c, width(), c);
+		p.drawLine(0, (int)c, width(), (int)c);
 		if(i>0)
-			p.drawText(2,c,KviQString::makeSizeReadable(m_maxRate * (KVI_IOGRAPH_HORIZ_SEGMENTS - i) / KVI_IOGRAPH_HORIZ_SEGMENTS));
+			p.drawText(2,(int)c,KviQString::makeSizeReadable(m_maxRate * (KVI_IOGRAPH_HORIZ_SEGMENTS - i) / KVI_IOGRAPH_HORIZ_SEGMENTS));
 		c+=sh;
 	}
 
 	c=1;
 	for(int i=0;i<=KVI_IOGRAPH_VERT_SEGMENTS;i++)
 	{
-		p.drawLine(c, 0, c, height());
+		p.drawLine((int)c, 0, (int)c, height());
 		c+=sw;
 	}
